Moves `removePunctuation` function into it's own utility module
