--- _includes/components/other-pages/SearchResult/partials/removePunctuation.js
@@ -1,3 +1,0 @@
-export default function removePunctuation(string) {
-  return string.replace(/[^\w|^\s]/g, ' ');
-}
