+++ _includes/utilities/js/helpers/removePunctuation.js
@@ -1,3 +1,3 @@
 export default function removePunctuation(string) {
-  return string.replace(/[^\w|^\s]/g, ' ');
+  return string.replace(/[^\w\s]/g, ' ');
 }
